Update generated models.

+++ deepwell/scripts/generate-models.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
 set -eu
 
+function join_arr() {
+	local IFS="$1"
+	shift
+	echo "$*"
+}
+
 # Only tables that are used are included
 #
 # Additionally, many Ozone tables have schemas
 # incompatible with SeaORM.
 
 tables=(
+	category
+	file
 	page
 	page_connection
 	page_connection_missing
@@ -16,6 +24,11 @@ tables=(
 	page_metadata
 	page_rate_vote
 	page_revision
+	settings
+	site
+	site_settings
+	user_block
+	user_messages
 	users
 )
 
@@ -23,10 +36,10 @@ tables=(
 cd "${0%/*}/.."
 
 # Delete old models
-rm -r src/models
+[[ -d src/models ]] && rm -r src/models
 
 # Generate models
 sea-orm-cli generate entity \
-	-t "${tables%,}" \
+	-t "$(join_arr , "${tables[@]}")" \
 	-u postgres://wikijump:wikijump@localhost/wikijump \
 	-o src/models
